# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sewa Barang API Documentation",
        "description": "Dokumentasi API untuk platform dashboard Sewa Barang",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/rent-transactions": {
            "get": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Daftar transaksi sewa",
                "description": "Mengambil semua riwayat transaksi sewa barang",
                "operationId": "getRentTransactionList",
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil data"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Buat transaksi sewa baru",
                "description": "Mendaftarkan transaksi sewa barang baru dengan QR Code",
                "operationId": "storeRentTransaction",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "qr_uuid",
                                    "region_id",
                                    "renter_name",
                                    "renter_phone",
                                    "rent_date",
                                    "expected_return_date",
                                    "qty",
                                    "rent_price"
                                ],
                                "properties": {
                                    "qr_uuid": {
                                        "type": "string",
                                        "example": "uuid-qr-123"
                                    },
                                    "region_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "renter_name": {
                                        "type": "string",
                                        "example": "Budi Utomo"
                                    },
                                    "renter_phone": {
                                        "type": "string",
                                        "example": "08123456789"
                                    },
                                    "rent_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-04-02"
                                    },
                                    "expected_return_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-04-05"
                                    },
                                    "qty": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "rent_price": {
                                        "type": "integer",
                                        "example": 100000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Transaksi berhasil dibuat"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/rent-transactions-paginated": {
            "get": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Daftar transaksi sewa (Paginated)",
                "description": "Mengambil daftar transaksi dengan sistem pagination",
                "operationId": "getRentTransactionPaginated",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Jumlah data per halaman",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil data"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/rent-transactions/{id}": {
            "put": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Update transaksi",
                "description": "Memperbarui data transaksi sewa berdasarkan ID",
                "operationId": "updateRentTransaction",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Transaksi",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "renter_name": {
                                        "type": "string",
                                        "example": "Budi Utomo Updated"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Transaksi berhasil diperbarui"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Hapus transaksi",
                "description": "Menghapus riwayat transaksi sewa berdasarkan ID",
                "operationId": "destroyRentTransaction",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Transaksi",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transaksi berhasil dihapus"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}